+++ how-to/workspace-platform-starter/package.json
@@ -8,9 +8,16 @@
 		"build": "npm run build-server & npm run build-client",
 		"build-server": "tsc --project ./server",
 		"prebuild-client": "node -p \"'// Generated from package.json version at build time. Do not modify directly.\\nexport const PLATFORM_VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > client/src/framework/platform/platform-version.ts",
-		"build-client": "npm run prebuild-client && npm run build-client-parallel",
-		"build-client-parallel": "node scripts/webpack-parallel.mjs ./client/webpack.config.js",
-		"build-client-serial": "webpack build --config ./client/webpack.config.js --mode=production",
+		"build-client": "npm run build-framework && npm run build-starter-modules && npm run build-client-modules",
+		"build-framework": "npm run prebuild-client && npm run build-framework-parallel",
+		"build-framework-parallel": "node scripts/webpack-parallel.mjs ./client/framework.webpack.config.js framework",
+		"build-framework-serial": "webpack build --config ./client/framework.webpack.config.js --mode=production",
+		"build-client-modules": "npm run build-client-modules-parallel",
+		"build-client-modules-parallel": "node scripts/webpack-parallel.mjs ./client/webpack.config.js client-modules",
+		"build-client-modules-serial": "webpack build --config ./client/webpack.config.js --mode=production",
+		"build-starter-modules": "npm run build-starter-modules-parallel",
+		"build-starter-modules-parallel": "node scripts/webpack-parallel.mjs ./client/starter-modules.webpack.config.js starter-modules",
+		"build-starter-modules-serial": "webpack build --config ./client/starter-modules.webpack.config.js --mode=production",
 		"build-templates": "webpack build --config ./templates/webpack.config.js --mode=production",
 		"start": "npm run server",
 		"client": "node ./scripts/launch.mjs",
